Create csc1 A2.txt
second assignment
/*
Create a class called Bank. It should have a variable for amount,
a method for deposit and a method for withdrawal.
There should also be an array of doubles that keeps track of every
deposit or withdrawal.  Create a method called listInteractions
that prints the array of deposits and withdrawals. The deposits
should be positive and the withdrawals should be negative.
*/